python-clients/scripts/asr/transcribe_file.py
import argparse
import grpc
import json
import os
import sys
import wave
from google.protobuf import duration_pb2

def transcribe_file(args):
    try:
        # Read audio file
        with wave.open(args.input_file, 'rb') as wav_file:
            audio_data = wav_file.readframes(wav_file.getnframes())
            sample_rate = wav_file.getframerate()
            num_channels = wav_file.getnchannels()

        # Set up gRPC channel
        if args.use_ssl:
            credentials = grpc.ssl_channel_credentials()
            channel = grpc.secure_channel(args.server, credentials)
        else:
            channel = grpc.insecure_channel(args.server)

        # Create metadata
        metadata = []
        if args.metadata:
            for meta in args.metadata:
                key, value = meta.split('=')
                metadata.append((key.strip(), value.strip()))

        # Create request
        request = {
            'audio': {
                'data': audio_data,
                'sample_rate': sample_rate,
                'num_channels': num_channels
            },
            'config': {
                'language_code': args.language_code
            }
        }

        # Make the request
        response = None  # This would be replaced with actual gRPC call
        
        # Print the response
        print(json.dumps(response, indent=2))

    except Exception as e:
        print(f"Error: {str(e)}", file=sys.stderr)
        sys.exit(1)

def main():
    parser = argparse.ArgumentParser(description='Transcribe audio file using NVIDIA NeMo ASR service')
    parser.add_argument('--server', required=True, help='Server address')
    parser.add_argument('--use-ssl', action='store_true', help='Use SSL for connection')
    parser.add_argument('--metadata', action='append', help='Metadata in key=value format')
    parser.add_argument('--language-code', default='en-US', help='Language code')
    parser.add_argument('--input-file', required=True, help='Path to input audio file')
    
    args = parser.parse_args()
    transcribe_file(args)

if __name__ == '__main__':
    main()